+++ apps/docliQ-mobile/src/components/ui/index.ts
@@ -1,2 +1,3 @@
 export * from './Button'
 export * from './ConfirmModal'
+export * from './GoogleGIcon'